+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.1.1] - 2026-08-10 - The turso 0.7.2 engine upgrade
 
 ### Changed